Apply spec compliant naming
Across the project, the term member was used to describe both instance
properties and instance methods in cases where only one of these terms
applied which could lead to ambiguity. Adding to this, usage of the
term 'member' is not widespread.

This commit changes member to property and instance method, respectively.