Improve custom converters tests
Now also tests if the custom converters work when specifying the type
in the decorator.

Also improved the test which checks whether a custom fromJson converter
influences the toJson of the type by constructing the test subject
manually instead of using parse which could result in two problems
canceling each other out.