Replace json dedector with separate parsing methods
`shouldOmitParsingString` was an ugly and inextensible function which
attempted to decide whether a string was supposed to be parsed before
being passed to the converter. This distinction cannot be made reliably
and to that end, the new parseJson method, which performs JSON
parsing, and parsePlain, which does not perform JSON parsing, are
introduced.