build(deps)!: upgrade TypeScript to ~4.9.5, was 3.9.7
Upgrade TypeScript to 4.9.5. This resolves the
`Type 'T' does not satisfy the constraint 'Object'.` errors when using
the library with TypeScript >= 4.8.0 and the
`Implicit conversion of a 'symbol' to a 'string' will fail at runtime.`
with TypeScript >= 4.7.0.

BREAKING CHANGE: Some generic types have been changed slightly in
regard to extending `Object`. This should not cause a problem.